+++ src/check-project/index.js
@@ -4,7 +4,6 @@ import fs from 'fs-extra'
 import path from 'path'
 import { execa } from 'execa'
 import prompt from 'prompt'
-import glob from 'it-glob'
 import { monorepoManifest } from './manifests/monorepo.js'
 import { typedESMManifest } from './manifests/typed-esm.js'
 import { typescriptManifest } from './manifests/typescript.js'
@@ -23,6 +22,13 @@ import {
 import semver from 'semver'
 import Listr from 'listr'
 import yargsParser from 'yargs-parser'
+import { fileURLToPath } from 'url'
+import {
+  isMonorepoProject,
+  glob
+} from '../utils.js'
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
 /**
  * @param {string} projectDir
@@ -125,6 +131,9 @@ async function processMonorepo (projectDir, manifest, branchName, repoUrl) {
   proposedManifest = sortManifest(proposedManifest)
 
   await ensureFileHasContents(projectDir, 'package.json', JSON.stringify(proposedManifest, null, 2))
+  await ensureFileHasContents(projectDir, '.gitignore', fs.readFileSync(path.join(__dirname, 'files', 'gitignore'), {
+    encoding: 'utf-8'
+  }))
   await checkLicenseFiles(projectDir)
   await checkBuildFiles(projectDir, branchName, repoUrl)
   await checkMonorepoReadme(projectDir, repoUrl, branchName, projectDirs)
@@ -382,6 +391,13 @@ async function processModule (projectDir, manifest, branchName, repoUrl, homePag
   proposedManifest = sortManifest(proposedManifest)
 
   await ensureFileHasContents(projectDir, 'package.json', JSON.stringify(proposedManifest, null, 2))
+
+  if (!isMonorepoProject(projectDir)) {
+    await ensureFileHasContents(projectDir, '.gitignore', fs.readFileSync(path.join(__dirname, 'files', 'gitignore'), {
+      encoding: 'utf-8'
+    }))
+  }
+
   await checkLicenseFiles(projectDir)
   await checkReadme(projectDir, repoUrl, branchName, rootManifest)
 }