+++ src/check-project/utils.js
@@ -60,7 +60,8 @@ export async function ensureFileHasContents (projectDir, filePath, expectedConte
     })
 
     if (!createFile) {
-      throw new Error(`Not creating ${filePath} file`)
+      console.warn(kleur.yellow(`${filePath} did not exist`))
+      return
     }
 
     fs.mkdirSync(path.dirname(path.join(projectDir, filePath)), {
@@ -106,7 +107,8 @@ export async function ensureFileHasContents (projectDir, filePath, expectedConte
     })
 
     if (!overwriteFile) {
-      throw new Error(`Not overwriting ${filePath} file`)
+      console.warn(kleur.yellow(`Not overwriting ${filePath} file`))
+      return
     }
 
     fs.mkdirSync(path.dirname(path.join(projectDir, filePath)), {